[refactor] : requestLine와 Header 객체 내부 클래스로 변경 - 외부에서 접근하는 동작이 아예 없으므로
