--- nameui/src/main/java/http/request/Header.java
@@ -1,16 +1,0 @@
-package http.request;
-
-import java.util.Map;
-
-public class Header {
-
-    private final Map<String, String> header;
-
-    public Header(Map<String, String> header) {
-        this.header = header;
-    }
-
-    public String getHeaderValue(String value) {
-        return header.get(value);
-    }
-}
