--- nameui/src/main/java/http/request/RequestLine.java
@@ -1,18 +1,0 @@
-package http.request;
-
-import constants.HttpMethod;
-import lombok.Getter;
-
-@Getter
-public class RequestLine {
-
-    private final HttpMethod method;
-    private final String uri;
-    private final String query;
-
-    public RequestLine(HttpMethod method, String uri, String query) {
-        this.method = method;
-        this.uri = uri;
-        this.query = query;
-    }
-}
